requirements.txt
openai-whisper
sounddevice
pyautogui
numpy
scipy
keyboard
